+++ src/user/user_api.h
@@ -213,7 +213,7 @@ MJAPI const char* mjs_setToAdhesion(mjsActuator* actuator, double gain);
 MJAPI const char* mjs_setToDCMotor(mjsActuator* actuator, double motorconst[2], double resistance,
                                    double nominal[3], double saturation[3], double inductance[2],
                                    double cogging[3], double controller[6], double thermal[6],
-                                   double lugre[5], int input_mode);
+                                   double lugre[5], int ctrlspec);
 
 
 //---------------------------------- Add assets ----------------------------------------------------