+++ unity/Runtime/Bindings/MjBindings.cs
@@ -305,6 +305,8 @@ public enum mjtCtrlInput : int{
   mjINPUT_POS = 1,
   mjINPUT_VEL = 2,
   mjINPUT_FF = 4,
+  mjINPUT_VOLTAGE = 8,
+  mjINPUT_NONE = 16,
 }
 public enum mjtObj : int{
   mjOBJ_UNKNOWN = 0,